Add support for sending status reports via Bluetooth.
This new feature allows Gyrid to send status update
reports to specified Bluetooth devices.

The MAC addresses of these devices have to be added to
to configuration file beforehand. Devices can receive
at most one report each hour of the day.

Device discovery is paused while setting up a connection
to send the report. One is advised to enable Bluetooth only
when trying to receive a report (and accept the connection
immediately when prompted).