Introduce networking support.
Gyrid now supports talking to a remote server over the network. This
can be both a local area network (LAN) or the internet (WAN).

The connection between the Gyrid client and the server runs over TCP
port 2583 by default and is tunnelled through SSL for security.

All data that is normally logged can be sent to the server, that means
informational messages as well as scan- and rssi-logs. The latter are
not sent by default but can be turned on by a server request.

Server hostname or IP-address can be configured in the configuration file.